fix: resolve CodeQL security alerts (log injection, path injection, code quality)
- Sanitize request URI before logging to prevent CRLF log injection (#407)
- Strengthen path validation in FileSystemController using realBasePath.resolve() (#406)
- Add path traversal validation in PesitReceiveService before file I/O (#405)
- Replace Lombok @Getter with explicit @Override methods in ClusterService (#408, #409)
- Cast int multiplication to long in RateLimitingFilter to prevent overflow (#403, #404)
- Add explicit NumberFormatException handling in ApiKey.ipToLong (#401)
- Remove user input echo from error responses (hardening)

+++ pesitwizard-server/src/main/java/com/pesitwizard/server/security/RateLimitingFilter.java
@@ -214,7 +214,7 @@ public class RateLimitingFilter extends OncePerRequestFilter {
         return Bucket.builder()
                 .addLimit(
                         Bandwidth.builder()
-                                .capacity(requestsPerMinute * burstMultiplier)
+                                .capacity((long) requestsPerMinute * burstMultiplier)
                                 .refillGreedy(requestsPerMinute, Duration.ofMinutes(1))
                                 .build())
                 .build();
@@ -231,7 +231,7 @@ public class RateLimitingFilter extends OncePerRequestFilter {
         return Bucket.builder()
                 .addLimit(
                         Bandwidth.builder()
-                                .capacity(requestsPerMinute * burstMultiplier)
+                                .capacity((long) requestsPerMinute * burstMultiplier)
                                 .refillGreedy(requestsPerMinute, Duration.ofMinutes(1))
                                 .build())
                 .build();
